Added a new preference, Trusted Hostname
If your remote database is hosted on a shared host and you're using HTTPS it's likely that the Certificate provided by your hosting provider will not have your website URL as the Common Name (it'll probably be some generic name shared by all domains hosted on that server). Android will reject the certificate on that basis. By entering your hostname in this field you're telling UPM to trust it.