Fixed a bug that was caused an NPE in DownloadRemoteDatabase.java.
I think the problem was that the temp file being written to (with File.createTempFile()) was being deleted by the system before I had a chance to operate on it. The temp file idea is unnecessary anyway since the bytes are already i memory. The fix involves using the in memory bytes rather than going though a temp file.