When AppEntryActivity.onCreate() is called check if savedInstanceState is null. If it is then display EnterMasterPassword.
This reason for this is to cater for the scenario where EnterMasterPassword.databaseFileToDecrypt is not null, i.e. when the app was already running and then was exited and then restarted. In these cases the AppEntryActivity wasn't moving on to another activity so was just leaving a blank screen.

savedInstanceState is null when the application is starting for the first time.