Fix crash problem when opening database
When opening a database the decryption happens in background task.
A progress dialog is presented during this process. If the screen
orientation changes while all this is happening a new
EnterMasterPassword activity is created. This means that when the
dismiss method is called on the progress dialog from within the
background task an exception is thrown.

java.lang.IllegalArgumentException: View not attached to window manager
EnterMasterPassword$DecryptDatabase.onPostExecute(EnterMasterPassword.java:142)

Reading this post, http://stackoverflow.com/a/2418763/219525, the
simple solution is to configure the activity so that Android
doesn't create a new activity on screen rotation.