Fix crash when screen rotated during password decryption
When the device is rotated the EnterMasterPassword activity
is destroyed and a new instance created. The problem with the
old code was that it didn't recreate the ProgressDialog.
Subsequently, when the Async task finished an error was raised
when it tried to dismiss the dialog.

This fix manages the progress dialog lifecycle and ensures it's
destroyed and recreated in paralell with it's parent activity.

Inspiration for this fix came from,
  http://stackoverflow.com/questions/3821423/background-task-progress-dialog-orientation-change-is-there-any-100-working/3821998#3821998