Reopen db on AddEditAccount activity
If the AddEditActivity is in the foreground, UPM is sent to the background,
UPM's process is terminated, UPM is brought back to the foreground, the
AddEditAccount was throwing an NPE because the database no longer existed.

This fix opens the EnterMasterPassword activity from within onResume()
if the db is null.