Use HTTP GET to download the db file and patch up multipart POST syntax
The proposed changes fix adrian/upm-android#19 issue, enabling the use
of HTTP GET method instead of static HTTP POST to download the database
file.

The use of LF (instead of CRLF) confuses CGI.pm if CGI.pm is used to
process HTTP uploads. The proposed changes enable processing of HTTP
uploads by a custom CGI script in addition to, or as a replacement of,
the current PHP scripts.